rename from Navbar.vue to NavbarComponents.vue
