+++ client/src/App.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import NavBar from './components/NavBar.vue'
+import NavbarComponents from './components/NavbarComponents.vue'
 import { RouterView } from 'vue-router'
 </script>
 
 <template>
   <header>
-    <NavBar />
+    <NavbarComponents />
   </header>
   <RouterView />
 </template>
